Add files via upload
The data for the private ltd. company has been analysed for 3 years(2017-2020). The beta value is 0.00652 which is less than 1 indicates, that the stock's price is less volatile than the overall market.
The overall market prices are compared to the particular company(TITAN).
Downloaded nifty 50 data from Yahoo finance, got the adjusted closing price of both the stocks
Calculated the Individual and general stock returns. Loaded the data into R console
Plotted the stocks graphically and calculated the beta and the expected rate of returns using CAPM (if R{=8 & Rm = 11)
Interpreted the results.